Dedicated internal pkg for sftp client
Added sftclient package to carry remote file retrieval operation
